requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
httpx
python-dotenv
pydantic
